requirements.txt
streamlit>=1.30
google-genai>=1.0
python-dotenv>=1.0
pandas>=2.0
